feat: Add logic for "stateless" agents (#984)
